requirements.txt
Flask>=2.0
neo4j>=4.4
python-dotenv>=0.19
pdfplumber>=0.11.6